ASD-50: move configuration-fetching methods from core helper into AmazonConfig
