+++ src/Payment/Gateway/Request/VoidRequestBuilder.php
@@ -19,7 +19,6 @@ namespace Amazon\Payment\Gateway\Request;
 use Magento\Payment\Gateway\Request\BuilderInterface;
 use Magento\Framework\App\ProductMetadata;
 use Amazon\Payment\Gateway\Helper\SubjectReader;
-use Amazon\Core\Helper\Data;
 use Magento\Sales\Api\OrderRepositoryInterface;
 
 class VoidRequestBuilder implements BuilderInterface
@@ -36,11 +35,6 @@ class VoidRequestBuilder implements BuilderInterface
     private $subjectReader;
 
     /**
-     * @var Data
-     */
-    private $coreHelper;
-
-    /**
      * @var OrderRepositoryInterface
      */
     private $orderRepository;
@@ -50,16 +44,13 @@ class VoidRequestBuilder implements BuilderInterface
      *
      * @param ProductMetadata          $productMetadata
      * @param SubjectReader            $subjectReader
-     * @param Data                     $coreHelper
      * @param OrderRepositoryInterface $orderRepository
      */
     public function __construct(
         ProductMetaData $productMetadata,
         SubjectReader $subjectReader,
-        Data $coreHelper,
         OrderRepositoryInterface $orderRepository
     ) {
-        $this->coreHelper = $coreHelper;
         $this->productMetaData = $productMetadata;
         $this->subjectReader = $subjectReader;
         $this->orderRepository = $orderRepository;
